Don't overwrite methods of an external object
`response.raw` is an urllib3 Response object. `response.close()` is a method on a requests Response object. I'm not sure why this was added in the first place, but this should not work.

It's possible that older versions of Requests and urllib3 had a bug around this, but using 2.9.0 and up, this line doesn't do anything useful and could be harmful.